moving things around a bit, more progress
