+++ packages/anonymous.js/src/prover/burn/burn.js
